hw2/src/requirements.txt
absl-py==0.8.0
astor==0.7.1
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
bleach==1.5.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
docopt==0.6.2
entrypoints==0.3
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
html5lib==0.9999999
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.20
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.0
nbformat==4.4.0
nltk==3.4.5
notebook==6.0.1
numpy==1.17.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pipreqs==0.4.9
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
Pygments==2.4.2
pygpu==0.7.6
pylint==2.3.1
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
sklearn-contrib-lightning==0.5.0
tensorboard==1.8.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
textblob==0.15.3
Theano==1.0.4
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wrapt==1.11.1
xgboost==0.90
yarg==0.1.9
